requirements.txt
fastapi==0.79.0
numpy==1.21.6
uvicorn==0.12.3
aiofiles==0.6.0
tensorflow-cpu==2.9.1
python-multipart
keras
Pillow